Addressed review comments and organized code into separate folders for specific functionality.
