--- APITestingPOC.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 25.0.1706.4
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APITestingPOC", "APITestingPOC\APITestingPOC.csproj", "{E1839E5C-ADF0-4AC6-A1E1-AA87ECF877A5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E1839E5C-ADF0-4AC6-A1E1-AA87ECF877A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1839E5C-ADF0-4AC6-A1E1-AA87ECF877A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1839E5C-ADF0-4AC6-A1E1-AA87ECF877A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E1839E5C-ADF0-4AC6-A1E1-AA87ECF877A5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1C53C4BD-3F55-4BFD-BC39-03B6DF71AE19}
-	EndGlobalSection
-EndGlobal
