--- APITestingPOC/APITestingPOC.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 25.0.1706.4
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APITestingPOC", "APITestingPOC.csproj", "{E8FE546D-3391-468E-9E1B-E1E026BFECE6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E8FE546D-3391-468E-9E1B-E1E026BFECE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E8FE546D-3391-468E-9E1B-E1E026BFECE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E8FE546D-3391-468E-9E1B-E1E026BFECE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E8FE546D-3391-468E-9E1B-E1E026BFECE6}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {AFE332B4-0653-497E-820B-6DA29B904388}
-	EndGlobalSection
-EndGlobal
