+++ APITestingPOC/Constants/APIConstants.cs
@@ -1,0 +1,5 @@
+﻿namespace APITestingPOC.Constants;
+    public class ApiConstants
+    {
+        public const string BaseUrl = "https://jsonplaceholder.typicode.com/";
+    }
