+++ APITestingPOC/Modals/GetUserResponse.cs
@@ -1,0 +1,8 @@
+﻿
+public class GetUserResponse
+{
+    public int userId { get; set; }
+    public int id { get; set; }
+    public string? title { get; set; }
+    public bool Completed { get; set; }
+}
