+++ APITestingPOC/Modals/PostUserResponse.cs
@@ -1,0 +1,7 @@
+﻿public class PostUserResponse
+{
+    public int id { get; set; }
+    public string? title { get; set; }
+    public string? body { get; set; }
+    public int userId { get; set; }
+}
