+++ APITestingPOC/Modals/PutUserResponse.cs
@@ -1,0 +1,6 @@
+﻿public class PutUserResponse
+{
+    public int id { get; set; }
+    public string? title { get; set; }
+    public bool completed { get; set; }
+}
