+++ APITestingPOC/Usings.cs
@@ -3,4 +3,5 @@ global using System.Net;
 global using Xunit;
 global using Xunit.Abstractions;
 global using RestSharp;
+global using Newtonsoft.Json;
 global using FluentAssertions;
