+++ APITestingPOC/Utils/TestSuiteUtil.cs
@@ -1,0 +1,12 @@
+﻿
+namespace APITestingPOC.Utils
+{
+	public class TestSuiteUtil
+	{
+        public static RestRequest CreateRestRequest(string resource, Method method)
+        {
+            return new RestRequest(resource, method);
+        }
+    }
+}
+
